refactor: simplify codebase by removing unnecessary null checks and casting operations throughout services and UI components.

+++ app/src/main/java/com/sameerasw/essentials/services/dreams/AmbientDreamService.kt
@@ -310,7 +310,7 @@ class AmbientDreamService : DreamService() {
                 FrameLayout.LayoutParams.MATCH_PARENT,
                 FrameLayout.LayoutParams.MATCH_PARENT
             )
-            setBackgroundColor(0x40000000.toInt())
+            setBackgroundColor(0x40000000)
         }
 
         clipContainer?.addView(imageView)
